Only attempt to use `org.freedesktop.ScreenSaver` if `gtk.Application.inhibit` fails and we're not on Flatpak
In general the GTK API should be good enough, but the fallback is probably a good idea for environments that
do not have the xdg-desktop-portal system installed. For the Flatpak however, the portal API should always
be available and the classic API will be blocked (due to sandboxing permissions), so it makes no sense to
compile the fallback on this platform.